main.dartのアプリ内表示部分をapp.dartに切り出し
